Modified SDFileSystem so it actually works.
Modified SDFileSystem to use the STM32F469NI discovery board SD library. This is because the library in it's original form hung forever when the class was constructed. I ripped out all the SD logic in it and replaced it with the STM32 HAL which does work.

Currently loads the background image and reports the size correctly in bytes.