Added bitmaps to GUI
Bitmaps weren't working with RAM they needed to be in flash memory. Because of this I've added a small codegen size. While doing this I found I ran out of memory so I moved to smaller tiles and tile them in code. Updated README to reflect this.

Next up buttons!